+++ stroom-query/stroom-query-api/src/main/java/stroom/query/api/v2/TableResultBuilder.java
@@ -14,7 +14,7 @@ public interface TableResultBuilder extends ResultBuilder<TableResultBuilder> {
 
     TableResultBuilder resultRange(OffsetRange resultRange);
 
-    TableResultBuilder totalResults(Integer totalResults);
+    TableResultBuilder totalResults(Long totalResults);
 
     TableResult build();
 }
